server: move serverpb to seperate file
